added configurationActivity; deleted start,registro1 y 2
