+++ app/src/main/java/com/example/administratec/GastoViewModel.kt
@@ -2,12 +2,15 @@ package com.example.administratec
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
+import java.util.*
 
 class GastoViewModel(private val gastoDao: GastosDao): ViewModel() {
 
     suspend fun obtenerGastos() = gastoDao.obtenerGastos()
 
     suspend fun obtenerGastosPorCategoria(categoria: String) = gastoDao.obtenerGastosPorCategoria(categoria)
+
+    suspend fun obtenerGastosPorFecha(inicio: Date, fin: Date) = gastoDao.obtenerGastosPorFecha(inicio,fin)
 
     suspend fun agregarGasto(gasto: Gasto) = gastoDao.agregarGasto(gasto)
