+++ app/src/main/java/com/example/administratec/GastosDao.kt
@@ -4,6 +4,7 @@ import androidx.room.Dao
 import androidx.room.Delete
 import androidx.room.Insert
 import androidx.room.Query
+import java.util.*
 
 @Dao
 interface GastosDao {
@@ -12,13 +13,10 @@ interface GastosDao {
 
     @Query("SELECT * FROM gasto WHERE categoria = :categoria")
     suspend fun obtenerGastosPorCategoria(categoria: String): List<Gasto>
-/*
-    @Query("SELECT * FROM gasto WHERE fecha")
-    suspend fun obtenerGatosSemanales(): List<Gasto>
 
-    @Query("SELECT * FROM gasto Where fecha")
-    suspend fun obtenerGastosMensuales(): List<Gasto>
-*/
+    @Query("SELECT * FROM gasto WHERE fecha BETWEEN :inicio AND :fin")
+    suspend fun obtenerGastosPorFecha(inicio: Date, fin: Date): List<Gasto>
+
     @Insert
     suspend fun agregarGasto(gasto: Gasto)
