--- app/src/main/java/com/example/administratec/StartActivity.kt
@@ -1,11 +1,0 @@
-package com.example.administratec
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-
-class StartActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_start)
-    }
-}
